Moved magefile repo root; created templates directory
