+++ README.md
@@ -1,80 +1,207 @@
-# goutils
+# goutils/v2/main
 
-[![License](https://img.shields.io/github/license/l50/goutils?label=License&style=flat&color=blue&logo=github)](https://github.com/l50/goutils/blob/main/LICENSE)
-[![Go Report Card](https://goreportcard.com/badge/github.com/l50/goutils/v2)](https://goreportcard.com/report/github.com/l50/goutils/v2)
-[![Tests](https://github.com/l50/goutils/actions/workflows/tests.yaml/badge.svg)](https://github.com/l50/goutils/actions/workflows/tests.yaml)
-[![🚨 CodeQL Analysis](https://github.com/l50/goutils/actions/workflows/codeql-analysis.yaml/badge.svg)](https://github.com/l50/goutils/actions/workflows/codeql-analysis.yaml)
-[![🚨 Semgrep Analysis](https://github.com/l50/goutils/actions/workflows/semgrep.yaml/badge.svg)](https://github.com/l50/goutils/actions/workflows/semgrep.yaml)
-[![Coverage Status](https://coveralls.io/repos/github/l50/goutils/badge.svg?branch=main)](https://coveralls.io/github/l50/goutils?branch=main)
-[![Renovate](https://github.com/l50/goutils/actions/workflows/renovate.yaml/badge.svg)](https://github.com/l50/goutils/actions/workflows/renovate.yaml)
-
-This repo is comprised of utilities that I use across various go projects.
-
-## Dependencies
-
-- [Install pre-commit](https://pre-commit.com/):
-
-  ```bash
-  python3 -m pip install --upgrade pip
-  python3 -m pip install pre-commit
-  ```
-
-- [Install Mage](https://magefile.org/):
-
-  ```bash
-  go install github.com/magefile/mage@latest
-  ```
-
-- [Install Keeper Commander](https://github.com/Keeper-Security/Commander):
-
-  ```bash
-  python3 -m pip install --upgrade pip
-  python3 -m pip install keepercommander
-  ```
+The `main` package is a collection of utility functions
+designed to simplify common main tasks.
 
 ---
 
-## For Contributors and Developers
+## Table of contents
 
-1. [Fork this project](https://docs.github.com/en/get-started/quickstart/fork-a-repo)
-
-1. Install dependencies:
-
-   ```bash
-   mage installDeps
-   ```
-
-1. Update and run pre-commit hooks locally:
-
-   ```bash
-   mage runPreCommit
-   ```
+- [Functions](#functions)
+- [Installation](#installation)
+- [Usage](#usage)
+- [Tests](#tests)
+- [Contributing](#contributing)
+- [License](#license)
 
 ---
 
-## Create New Release
+## Functions
 
-- Download and install the [gh cli tool](https://cli.github.com/):
+### FixCodeBlocks(string, string)
 
-  - [macOS](https://github.com/cli/cli#macos)
-  - [Linux](https://github.com/cli/cli/blob/trunk/docs/install_linux.md)
-  - [Windows](https://github.com/cli/cli#windows)
+```go
+FixCodeBlocks(string, string) error
+```
 
-- Install changelog extension:
+FixCodeBlocks fixes code blocks for the input filepath
+using the input language.
 
-  ```bash
-  gh extension install chelnak/gh-changelog
-  ```
+Example usage:
 
-- Generate changelog:
+```go
+mage fixcodeblocks go docs/docGeneration.go
+```
 
-  ```bash
-  NEXT_VERSION=v1.1.3
-  gh changelog new --next-version "${NEXT_VERSION}"
-  ```
+**Parameters:**
 
-- Create release:
+filepath: the path to the file or directory to fix
 
-  ```bash
-  gh release create "${NEXT_VERSION}" -F CHANGELOG.md
-  ```
+language: the language of the code blocks to fix
+
+**Returns:**
+
+error: an error if one occurred
+
+---
+
+### GeneratePackageDocs()
+
+```go
+GeneratePackageDocs() error
+```
+
+GeneratePackageDocs creates documentation for the various packages
+in the project.
+
+Example usage:
+
+```go
+mage generatepackagedocs
+```
+
+**Returns:**
+
+error: An error if any issue occurs during documentation generation.
+
+---
+
+### InstallDeps()
+
+```go
+InstallDeps() error
+```
+
+InstallDeps installs the Go dependencies necessary for developing
+on the project.
+
+Example usage:
+
+```go
+mage installdeps
+```
+
+**Returns:**
+
+error: An error if any issue occurs while trying to
+install the dependencies.
+
+---
+
+### RunPreCommit()
+
+```go
+RunPreCommit() error
+```
+
+RunPreCommit updates, clears, and executes all pre-commit hooks
+locally. The function follows a three-step process:
+
+First, it updates the pre-commit hooks.
+Next, it clears the pre-commit cache to ensure a clean environment.
+Lastly, it executes all pre-commit hooks locally.
+
+Example usage:
+
+```go
+mage runprecommit
+```
+
+**Returns:**
+
+error: An error if any issue occurs at any of the three stages
+of the process.
+
+---
+
+### RunTests()
+
+```go
+RunTests() error
+```
+
+RunTests executes all unit tests.
+
+Example usage:
+
+```go
+mage runtests
+```
+
+**Returns:**
+
+error: An error if any issue occurs while running the tests.
+
+---
+
+### UpdateMirror(string)
+
+```go
+UpdateMirror(string) error
+```
+
+UpdateMirror updates pkg.go.dev with the release associated with the
+input tag
+
+Example usage:
+
+```go
+mage updatemirror v2.0.1
+```
+
+**Parameters:**
+
+tag: the tag to update pkg.go.dev with
+
+**Returns:**
+
+error: An error if any issue occurs while updating pkg.go.dev
+
+---
+
+## Installation
+
+To use the goutils/v2/main package, you first need to install it.
+Follow the steps below to install via go get.
+
+```bash
+go get github.com/l50/goutils/v2/main
+```
+
+---
+
+## Usage
+
+After installation, you can import the package in your Go project
+using the following import statement:
+
+```go
+import "github.com/l50/goutils/v2/main"
+```
+
+---
+
+## Tests
+
+To ensure the package is working correctly, run the following
+command to execute the tests for `goutils/v2/main`:
+
+```bash
+go test -v
+```
+
+---
+
+## Contributing
+
+Pull requests are welcome. For major changes,
+please open an issue first to discuss what
+you would like to change.
+
+---
+
+## License
+
+This project is licensed under the MIT
+License - see the [LICENSE](../LICENSE)
+file for details.
