+++ magefile.go
@@ -4,10 +4,13 @@
 package main
 
 import (
+	"bufio"
 	"fmt"
+	"io"
 	"log/slog"
 	"os"
 	"path/filepath"
+	"strings"
 
 	"github.com/fatih/color"
 	"github.com/l50/goutils/v2/dev/lint"
@@ -96,7 +99,7 @@ func GeneratePackageDocs() error {
 		Name:  "goutils/v2",
 	}
 
-	templatePath := filepath.Join("magefiles", "tmpl", "README.md.tmpl")
+	templatePath := filepath.Join("templates", "README.md.tmpl")
 	if err := docs.CreatePackageDocs(fs, repo, templatePath); err != nil {
 		return fmt.Errorf("failed to create package docs: %v", err)
 	}
@@ -232,4 +235,70 @@ func TestLoggerOutput() {
 	log.Debugf("This is a test debug message")
 	log.Errorf("This is a test %s error message", "formatted")
 	log.Println("{\"time\":\"2024-01-03T23:12:35.937476-07:00\",\"level\":\"ERROR\",\"msg\":\"\\u001b[1;32m==> docker.ansible-attack-box: Starting docker container...\\u001b[0m\"}")
+}
+
+// RunTests executes all unit tests.
+//
+// Example usage:
+//
+// ```go
+// mage runtests
+// ```
+//
+// **Returns:**
+//
+// error: An error if any issue occurs while running the tests.
+func RunTests() error {
+	fmt.Println("Running unit tests.")
+	if _, err := sys.RunCommand(filepath.Join(".hooks", "run-go-tests.sh"), "all"); err != nil {
+		return fmt.Errorf("failed to run unit tests: %v", err)
+	}
+
+	return nil
+}
+
+// processLines parses an io.Reader, identifying and marking code blocks
+// found in a README.
+func processLines(r io.Reader, language string) ([]string, error) {
+	scanner := bufio.NewScanner(r)
+	var lines, codeBlockLines []string
+	var inCodeBlock bool
+
+	for scanner.Scan() {
+		line := scanner.Text()
+
+		inCodeBlock, codeBlockLines = handleLineInCodeBlock(strings.TrimSpace(line), line, inCodeBlock, language, codeBlockLines)
+
+		if !inCodeBlock {
+			lines = append(lines, codeBlockLines...)
+			codeBlockLines = codeBlockLines[:0]
+			if !strings.HasPrefix(line, "```") {
+				lines = append(lines, line)
+			}
+		}
+	}
+
+	if inCodeBlock {
+		codeBlockLines = append(codeBlockLines, "\t\t\t// ```")
+		lines = append(lines, codeBlockLines...)
+	}
+
+	return lines, scanner.Err()
+}
+
+// handleLineInCodeBlock categorizes and handles each line based on its
+// content and relation to code blocks found in a README.
+func handleLineInCodeBlock(trimmedLine, line string, inCodeBlock bool, language string, codeBlockLines []string) (bool, []string) {
+	switch {
+	case strings.HasPrefix(trimmedLine, "```"+language):
+		if !inCodeBlock {
+			codeBlockLines = append(codeBlockLines, line)
+		}
+		return !inCodeBlock, codeBlockLines
+	case inCodeBlock:
+		codeBlockLines = append(codeBlockLines, line)
+	case strings.Contains(trimmedLine, "```"):
+		inCodeBlock = false
+	}
+	return inCodeBlock, codeBlockLines
 }
